+++ saleor/graphql/shop/types.py
@@ -114,7 +114,12 @@ class Shop(graphene.ObjectType):
     description = graphene.String(description="Shop's description.")
     domain = graphene.Field(Domain, required=True, description="Shop's domain data.")
     homepage_collection = graphene.Field(
-        Collection, description="Collection displayed on homepage."
+        Collection,
+        description="Collection displayed on homepage.",
+        deprecation_reason=(
+            "Use the `collection` query with the `slug` parameter. "
+            "This field will be removed in Saleor 3.0"
+        ),
     )
     languages = graphene.List(
         LanguageDisplay,
@@ -122,7 +127,11 @@ class Shop(graphene.ObjectType):
         required=True,
     )
     name = graphene.String(description="Shop's name.", required=True)
-    navigation = graphene.Field(Navigation, description="Shop's navigation.")
+    navigation = graphene.Field(
+        Navigation,
+        description="Shop's navigation.",
+        deprecation_reason="Fetch menus using the `menu` query with `slug` parameter.",
+    )
     permissions = graphene.List(
         Permission, description="List of available permissions.", required=True
     )