+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Allow product variants to be sortable - #6138 by @tomaszszymanski129
 - Staff with only manage_orders should be able to query all stock related date. - #6139 by @fowczarek
 - Add filtering to `ProductVariants` query and option to fetch variant by sku in `ProductVariant` query - #6190 by @fowczarek
+- Add filtering by Product IDs to `products` query - #6224 by @GrzegorzDerdak
 
 
 ### Breaking Changes
@@ -34,6 +35,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add dummy credit card payment - #5822 by @IKarbowiak
 - Anonymize and update order and payment fields; drop PaymentSecureConfirm mutation, drop Payment type fields: extraData, billingAddress, billingEmail, drop gatewayResponse from Transaction type - #5926 by @IKarbowiak
 - Switch the HTTP stack from WSGI to ASGI based on Uvicorn - #5960 by @patrys
+- Add manage product types and attributes permission - #6219 by @IKarbowiak
 
 ### Fixes
 
@@ -69,6 +71,9 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix get unpublished products, product variants and collection as app - #6194 by @fowczarek
 - Set OrderFulfillStockInput fields as required - #6196 by @IKarbowiak
 - Fix attribute filtering by categories and collections - #6214 by @fowczarek
+- Fix is_visible when publication_date==today - #6225 by @korycins
+- Fix filtering products by multiple attributes - #6215 by @GrzegorzDerdak
+- Add attributes validation while creating/updating a product's variant - #6269 by @GrzegorzDerdak
 
 ## 2.10.2